+++ sdk/resourcemanager/healthcareapis/armhealthcareapis/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release History
 
+## 0.5.0 (2022-05-02)
+### Features Added
+
+- New function `CorsConfiguration.MarshalJSON() ([]byte, error)`
+- New struct `CorsConfiguration`
+- New field `CorsConfiguration` in struct `DicomServiceProperties`
+
+
 ## 0.4.0 (2022-04-15)
 ### Breaking Changes